Add files via upload
The goal is to read data from two files, "ShapeStrokeData.txt" and "DrawShapes.txt," which contain information on shapes and drawing instructions respectively. 
The software should read the data from these files, process it, and send commands in the form of G-Code to the Arduino to control the drawing arm and produce drawings on a grid. The size of the shapes should be 20mm high. 
The user should be able to input the filename for the drawing instructions. The software should be able to handle multiple sets of shapes from different instruction files in the same grid. The pen should finish in the pen-up position.